refactor(client): rename exchange rates widget file
- Rename exchange-rates.tsx to exchange-rates-widget.tsx for consistent naming convention with other widget files
