+++ fullstack/task/packages/client/src/widgets/exchange-rates-table/ui/exchange-rates.tsx
