+++ fullstack/task/packages/client/src/widgets/exchange-rates-table/ui/index.ts
@@ -1,1 +1,1 @@
-export { ExchangeRates } from './ExchangeRates';
+export { ExchangeRates } from './exchange-rates';
